Optimized viewer, fixed type output problem (all float precision now).

+++ Particle.cpp
@@ -2,9 +2,9 @@
 
 Particle::Particle () : mass(1), radius(0.1), position(), velocity(), angle(0), omega() {}
 
-Particle::Particle (double m, double r) : mass(m), radius(r), position(), velocity(), angle(0), omega() {};
+Particle::Particle (float m, float r) : mass(m), radius(r), position(), velocity(), angle(0), omega() {};
 
-Particle::Particle (double m, double r, Vector3 p, Vector3 v, double a, Vector3 o) : mass(m), radius(r), position(p.x, p.y, p.z), velocity(v.x, v.y, v.z), angle(a), omega(o.x, o.y, o.z) {};
+Particle::Particle (float m, float r, Vector3 p, Vector3 v, float a, Vector3 o) : mass(m), radius(r), position(p.x, p.y, p.z), velocity(v.x, v.y, v.z), angle(a), omega(o.x, o.y, o.z) {};
 
 Particle::~Particle () {}
 
@@ -16,17 +16,17 @@ void Particle::initialize () {
 	radius = random(0.1, 0.2);
 }
 
-void Particle::positionStep (double dt, Vector3& force) {
+void Particle::positionStep (float dt, Vector3& force) {
 
 	position = position + velocity * dt;
 }
 
-void Particle::velocityStep (double dt, Vector3& force) {
+void Particle::velocityStep (float dt, Vector3& force) {
 	
 	velocity = velocity + (force/mass) * dt;
 }
 
-void Particle::angularStep (double dt) {
+void Particle::angularStep (float dt) {
 	
 	angle = fmod(angle + omega.norm() * dt, 2*PI);
 }
